seperated the methods into different file.
added multiple breadcrumb support
List/Grid bug fix
added source build for debugging
spellcheck going back issue
