+++ package.json
@@ -16,7 +16,8 @@
     "build:prod": "webpack --config ./webpack/webpack.prod.config.js",
     "server": "http-server ./public",
     "generate-docs": "jsdoc -c ./docs/jsDoc.json ",
-    "upload-s3":"node s3Upload.js"
+    "upload-s3":"node s3Upload.js",
+    "build:dev":"webpack ---config ./webpack/webpack.dev.build.js"
   },
   "devDependencies": {
     "@babel/core": "^7.10.5",
